Update index.js
this is one of the best pages I have made so far using JS. it was very fun coding this game website. so if you ever want to contact me then you know where to find me. and if you have a job to offer then you know where to find me to